refactor(providers): extract AQE-router pipeline into aqe-router.mjs
providers.mjs (1,086 counted lines, over the max-lines warning) hosted
the whole applyAqeRouter/aqeRouterDrift/undoAqeRouter convergence
pipeline: AQE_ROUTER_SURFACES, the four surface reconcilers,
buildAqeRouterContext/runAqeRouterFold, and reconcileExternalProviders
(CC 40). All of it moves to a new src/lib/aqe-router.mjs, which imports
back the shared hash/ownership primitives (AQE_OWNERSHIP_KEY,
stableValue, declarationHash, plainRecord — now exported) and the
provider-detection helpers it needs (aqeExternalProviders,
aqeRouterFile, aqeSupportsExternalProviders, aqeSupportsAgentOverrides,
aqeSelectableChainProviderTypes, credentialGaps) from providers.mjs.
providers.mjs re-imports applyAqeRouter/aqeRouterDrift/undoAqeRouter
and re-exports them, so every existing `./providers.mjs` import path
(commands, status sections, tests) keeps working unchanged — the two
modules do have a real bidirectional dependency (mirroring the
pre-existing, safe usage-index.mjs/usage-opencode.mjs pattern: hoisted
function declarations, nothing read at module-eval time), which is a
deliberate tradeoff against pulling much more of providers.mjs's public
detection API into the new file.

reconcileExternalProviders (CC 40) decomposes into
reconcileDesiredProvider/retireStaleProvider (with
reconcileConflictingDeclaration/reconcileProviderActivation) — each
well under CC 25. No behavior change: identical router-file writes,
identical drift detection, identical ownership receipts.
